[PT FE] Resolve i64 usage in frontend (#23518)
### Details:
 - *Support lightglue*
 - *Support int64 as scalar in PyTorch Frontend*

### Tickets:
 - *CVS-127326, CVS-134328*

+++ src/frontends/pytorch/src/op/len.cpp
@@ -21,7 +21,7 @@ OutputVector translate_len(const NodeContext& context) {
     auto const_0 = context.mark_node(v0::Constant::create(element::i32, Shape{1}, {0}));
     auto const_1 = context.mark_node(v0::Constant::create(element::i32, Shape{1}, {1}));
     auto input = context.get_input(0);
-    auto input_shape = context.mark_node(std::make_shared<v3::ShapeOf>(input, element::i32));
+    auto input_shape = context.mark_node(std::make_shared<v3::ShapeOf>(input, element::i64));
 
     auto slice = context.mark_node(std::make_shared<v8::Slice>(input_shape, const_0, const_1, const_1));
     // Slice will return empty tensor for empty lists, we use the fact that ReduceSum(empty tensor) = 0
